Implemented RigidBody::load() to parse an input file defining a RigidBody
ALSO: redefined methods addCircle() and addLine() to specify the approximate grid spacing as the final argument, instead of the number of points.
Retained the old interface (with integer number of points as the last argument) as addCircle_n() and addLine_n().  Could also overload these, so that the two versions have the same name, but this seems more likely to lead to bugs (e.g. 3 vs 3.0).


git-svn-id: https://svn.princeton.edu/rowley/trunk@121 f381375e-2823-4fb2-bad6-22fa55754e8c