Change forces written to file
Change forces written to file so that they are the same as the forces
output to the screen. This change makes it so that forces are
normalized by a constant dynamic pressure instead of a time-varying
dynamic pressure. For a plunging body, the forces should be normalized
by the dynamic pressure based on the freestream velocity, not based
on the instantaneous velocity seen by the body.