Add StateVector class
The StateVector class is like the State class, but allows for addition
and scalar multiplication of StateVector objects, hence the name. This
functionality is useful for numerical root-finding of a flow map, for
example.

State objects contain timestep and time member variables. It is unclear
how to best handle these variables when adding States or multiplying
them by a scalar. We have made the arbitrary choice of setting these
variables equal to zero for all StateVector objects; addition and
scalar multiplication do not change this.